refactor: update filename and move `DeleteSelectButton` to main components folder
