+++ src/components/index.ts
@@ -4,3 +4,4 @@ export * from "./QueryClientContextProvider";
 export * from "./Layout";
 export * from "./SearchBar";
 export * from "./AdvanceSearch";
+export * from "./DeleteSelectButton";
