+++ src/modules/lobby-serksa/components/data-tables/SchoolDataTable.tsx
@@ -1,8 +1,10 @@
 "use client";
 
-import CustomPagination from "@components/CustomPagination";
-import { quizColumns } from "@modules/lobby-serksa/components/data-tables";
-import { Button } from "@components/ui/button";
+import {
+  flexRender,
+  getCoreRowModel,
+  useReactTable,
+} from "@tanstack/react-table";
 import {
   Table,
   TableBody,
@@ -10,28 +12,26 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from "@components/ui/table";
-import {
-  flexRender,
-  getCoreRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-import { useQuery } from "@tanstack/react-query";
-import React from "react";
-import { parseAsInteger, parseAsString, useQueryState } from "nuqs";
-import { axiosInstance } from "@modules/lobby-serksa/fetch";
+} from "@/components/ui/table";
+import CustomPagination from "@components/CustomPagination";
 import { usePagination } from "@hooks";
-import Link from "next/link";
+import { Button } from "@components/ui/button";
+import { parseAsInteger, parseAsString, useQueryState } from "nuqs";
+import { useQuery } from "@tanstack/react-query";
+import { schoolColumns } from "@modules/lobby-serksa/components/data-tables";
+import { axiosInstance } from "@modules/lobby-serksa/fetch";
+import React from "react";
 import { SearchBar } from "@components";
-import { DeleteSelectButton } from "../DeleteSelectButton";
-import { QuizAdvanceSearch } from "../advance-searchs";
+import { SchoolCreateSheet } from "../create-sheets";
+import { SchoolAdvanceSearch } from "../advance-searchs";
+import { DeleteSelectButton } from "@components";
 
 function onDeleteSelectedHandler(ids: number[]) {
   const deleteBody = ids.map((id) => ({ id }));
-  return axiosInstance.delete(`/admin/quizzes`, { data: deleteBody });
+  return axiosInstance.delete(`/admin/schools`, { data: deleteBody });
 }
 
-export function QuizDataTable() {
+export function SchoolDataTable() {
   const [limit, setLimit] = useQueryState(
     "limit",
     parseAsInteger.withDefault(10),
@@ -41,70 +41,21 @@ export function QuizDataTable() {
     parseAsInteger.withDefault(0),
   );
   const [search, setSearch] = useQueryState(
-    "question:ilike",
+    "name:ilike",
     parseAsString.withDefault(""),
   );
   const [searchId, setSearchId] = useQueryState("id:eq", parseAsInteger);
-  const [searchArchived, setSearchArchived] = useQueryState(
-    "archived:eq",
-    parseAsString,
-  );
-  const [searchProfessorId, setSearchProfessorId] = useQueryState(
-    "base-professor-id:eq",
-    parseAsInteger,
-  );
-  const [searchSubjectId, setSearchSubjectId] = useQueryState(
-    "base-subject-id:eq",
-    parseAsInteger,
-  );
-  const [searchSemester, setSearchSemester] = useQueryState(
-    "semester:eq",
-    parseAsInteger,
-  );
-  const [searchYear, setSearchYear] = useQueryState("year:eq", parseAsInteger);
-  const [searchMajorId, setSearchMajorId] = useQueryState(
-    "major-id:eq",
-    parseAsInteger,
-  );
-  const [searchSchoolId, setSearchSchoolId] = useQueryState(
-    "school-id:eq",
-    parseAsInteger,
-  );
   const [total, setTotal] = React.useState(0);
   const [rowSelection, setRowSelection] = React.useState({});
 
   const { data, refetch, isLoading } = useQuery({
     queryKey: [
-      "quizzes",
-      {
-        limit,
-        offset,
-        "question:ilike": search,
-        "id:eq": searchId,
-        "archived:eq": searchArchived,
-        "base-professor-id:eq": searchProfessorId,
-        "base-subject-id:eq": searchSubjectId,
-        "semester:eq": searchSemester,
-        "year:eq": searchYear,
-        "major-id:eq": searchMajorId,
-        "school-id:eq": searchSchoolId,
-      },
+      "schools",
+      { limit, offset, "name:ilike": search, "id:eq": searchId },
     ],
     queryFn: async () => {
-      const response = await axiosInstance.get(`/admin/quizzes`, {
-        params: {
-          limit,
-          offset,
-          "question:ilike": search,
-          "id:eq": searchId,
-          "archived:eq": searchArchived,
-          "base-professor-id:eq": searchProfessorId,
-          "base-subject-id:eq": searchSubjectId,
-          "semester:eq": searchSemester,
-          "year:eq": searchYear,
-          "major-id:eq": searchMajorId,
-          "school-id:eq": searchSchoolId,
-        },
+      const response = await axiosInstance.get(`/admin/schools`, {
+        params: { limit, offset, "name:ilike": search, "id:eq": searchId },
       });
       setTotal(+response.headers["x-total-count"] || 0);
       return response.data;
@@ -112,8 +63,8 @@ export function QuizDataTable() {
   });
 
   const table = useReactTable({
-    data: data?.data.quizzes || [],
-    columns: quizColumns,
+    data: data?.data.schools || [],
+    columns: schoolColumns,
     getCoreRowModel: getCoreRowModel(),
     onRowSelectionChange: setRowSelection,
     state: {
@@ -143,38 +94,13 @@ export function QuizDataTable() {
               setSearch(searchPharse);
               go(1);
             }}
-            placeholder="Search question ..."
+            placeholder="Search name ..."
             defaultValue={search}
           />
-          <QuizAdvanceSearch
-            defaultValues={{
-              id: searchId,
-              archived: searchArchived,
-              professorId: searchProfessorId,
-              subjectId: searchSubjectId,
-              semester: searchSemester,
-              year: searchYear,
-              majorId: searchMajorId,
-              schoolId: searchSchoolId,
-            }}
-            onConfirm={({
-              id,
-              archived,
-              professorId,
-              subjectId,
-              semester,
-              year,
-              majorId,
-              schoolId,
-            }) => {
+          <SchoolAdvanceSearch
+            defaultValues={{ id: searchId }}
+            onConfirm={({ id }) => {
               setSearchId(id);
-              setSearchArchived(archived);
-              setSearchProfessorId(professorId);
-              setSearchSubjectId(subjectId);
-              setSearchSemester(semester);
-              setSearchYear(year);
-              setSearchMajorId(majorId);
-              setSearchSchoolId(schoolId);
               go(1);
             }}
           />
@@ -187,9 +113,9 @@ export function QuizDataTable() {
           refetch={refetch}
           table={table}
         />
-        <Link href="/lobby-serksa/quizzes/create">
+        <SchoolCreateSheet refetch={() => refetch()}>
           <Button>Create</Button>
-        </Link>
+        </SchoolCreateSheet>
       </div>
       <div className="overflow-hidden rounded-lg border">
         <Table>
@@ -231,7 +157,7 @@ export function QuizDataTable() {
             ) : (
               <TableRow>
                 <TableCell
-                  colSpan={quizColumns.length}
+                  colSpan={schoolColumns.length}
                   className="h-24 text-center"
                 >
                   {isLoading ? (
