+++ src/modules/lobby-serksa/components/data-tables/index.ts
@@ -1,10 +1,10 @@
 export * from "./school-columns";
-export * from "./school-data-table";
+export * from "./SchoolDataTable";
 export * from "./major-columns";
-export * from "./major-data-table";
+export * from "./MajorDataTable";
 export * from "./subject-columns";
-export * from "./subject-data-table";
+export * from "./SubjectDataTable";
 export * from "./professor-columns";
-export * from "./professor-data-table";
+export * from "./ProfessorDataTable";
 export * from "./quiz-columns";
-export * from "./quiz-data-table";
+export * from "./QuizDataTable";
