+++ src/modules/lobby-serksa/components/index.ts
@@ -1,3 +1,2 @@
 export * from "./SearchPopover";
-export * from "./DeleteSelectButton";
 export * from "./QuizImport";
